feat(LoginAction): Add auto user confirmation
This commit adds the functionality to auto confirm the user if the user doesn't exist already. It first checks if the user may exist in the database and if not, it creates a new user and confirms the email. If there's an error, it returns the error message.

fix(Database types): Add enums for ebook and order status

This commit adds the enums for ebook status and order status in the database types. The ebook status can be either "UNCHECKED" or "CHECKED" while the order status can be "CREATED", "ACCEPTED", "DELIVERED", "DONE", or "CANCELED".